fix: use redirectTo function for next router redirects
